+++ Sources/csuper-gui/csuper-gui.h
@@ -61,7 +61,7 @@ typedef struct
     GtkBuilder *ptr_builder;                /*!< A pointer to a GTK builder. */
     csuStruct *ptr_csu_struct;              /*!< A pointer to a csuStruct */
     GtkWidget *ptr_main_window;                 /*! The main window */
-    game_config config;                     /*!< A game configuration */
+    //game_config config;                     /*!< A game configuration */
     gchar csu_filename[SIZE_MAX_FILE_NAME];   /*!< The filename of the csu file*/
     GtkClipboard *ptr_clipboard;                /*!< A pointer to the main clipboard */
     GtkClipboard *ptr_clipboard_selected;       /*!< A pointer to the selected clipboard */