+++ Sources/csuper-gui/game_configuration.c
@@ -47,7 +47,7 @@ G_MODULE_EXPORT void openGameConfigurationPreferences(GtkWidget *widget, gpointe
     if (!window_game)
         g_critical(_("Widget game_configuration_preferences is missing in file csuper-gui.glade."));
 
-    displayGameConfiguration(window_game);
+    displayGameConfiguration(user_data);
 
     gtk_widget_show_all(window_game);
 }
@@ -66,9 +66,22 @@ G_MODULE_EXPORT void closeGameConfigurationPreferences(GtkWidget *widget, gpoint
     if (!window_game)
         g_critical(_("Widget game_configuration_preferences is missing in file csuper-gui.glade."));
 
-    deleteDisplayGameConfiguration(window_game);
+    deleteDisplayGameConfiguration(user_data);
 
     gtk_widget_hide(window_game);
+}
+
+/*!
+ * \fn G_MODULE_EXPORT gboolean closeGameConfigurationPreferencesQuit(GtkWidget *widget, GdkEvent  *event, gpointer user_data)
+ *  Close the game configuration preferences
+ * \param[in] widget the widget which send the interrupt
+ * \param[in] event the event which triggered this signal
+ * \param[in] data the globalData
+ */
+G_MODULE_EXPORT gboolean closeGameConfigurationPreferencesQuit(GtkWidget *widget, GdkEvent  *event, gpointer user_data)
+{
+    closeGameConfigurationPreferences(widget,user_data);
+    return TRUE;
 }
 
 /*!
@@ -182,6 +195,8 @@ G_MODULE_EXPORT void chooseImportedFile(GtkWidget *widget, gpointer data)
 	gtk_widget_destroy(window_file_import);
 	if (error)
         importGameConfigurationError(user_data);
+
+    updateDisplayGameConfiguration(user_data);
 }
 
 /*!
@@ -200,16 +215,18 @@ void importGameConfigurationError(globalData *data)
 }
 
 /*!
- * \fn void void displayGameConfiguration(GtkWidget *window_game)
+ * \fn void displayGameConfiguration(globalData *data)
  *  Display a all the game configuration on the window
- * \param[in] window_game the window of game configuration
+ * \param[in] data the globalData
  */
-void displayGameConfiguration(GtkWidget *window_game)
+void displayGameConfiguration(globalData *data)
 {
     gint i;
 
     /* Get the windows scrolled*/
-    GtkWidget *window_scrolled = gtk_grid_get_child_at(GTK_GRID(gtk_bin_get_child(GTK_BIN(window_game))),0,1);
+    GtkWidget *window_scrolled = GTK_WIDGET(gtk_builder_get_object(data->ptr_builder,"scrolled_window_game_configuration_preferences"));
+    if (!window_scrolled)
+        g_critical(_("Widget scrolled_window_game_configuration_preferences is missing in file csuper-gui.glade."));
 
     /*Read the game configurations*/
     list_game_config *ptr_list_config;
@@ -231,11 +248,27 @@ void displayGameConfiguration(GtkWidget *window_game)
     /* Fill the grid */
     for (i=0 ; i<ptr_list_config->nb_config ; i++)
     {
-        gtk_grid_attach(GTK_GRID(grid),gtk_label_new(ptr_list_config->name_game_config[i]),1,i,1,1);
-        gtk_grid_attach(GTK_GRID(grid),gtk_button_new_with_label("See details"),2,i,1,1);
+        gtk_grid_attach(GTK_GRID(grid),gtk_label_new(ptr_list_config->name_game_config[i]),0,i,1,1);
+        gtk_widget_set_hexpand(gtk_grid_get_child_at(GTK_GRID(grid),0,i),TRUE);
+        gtk_label_set_selectable(GTK_LABEL(gtk_grid_get_child_at(GTK_GRID(grid),0,i)),TRUE);
+        #ifdef _WIN32
+        gtk_grid_attach(GTK_GRID(grid),gtk_button_new_from_stock("gtk-edit"),1,i,1,1);
+        gtk_grid_attach(GTK_GRID(grid),gtk_button_new_from_stock("gtk-properties"),2,i,1,1);
         gtk_grid_attach(GTK_GRID(grid),gtk_button_new_from_stock("gtk-delete"),3,i,1,1);
-        gtk_widget_set_hexpand(gtk_grid_get_child_at(GTK_GRID(grid),1,i),TRUE);
-        gtk_label_set_selectable(GTK_LABEL(gtk_grid_get_child_at(GTK_GRID(grid),1,i)),TRUE);
+        #else
+        gtk_grid_attach(GTK_GRID(grid),gtk_button_new_from_icon_name("gtk-edit",GTK_ICON_SIZE_BUTTON),1,i,1,1);
+        gtk_grid_attach(GTK_GRID(grid),gtk_button_new_from_icon_name("gtk-properties",GTK_ICON_SIZE_BUTTON),2,i,1,1);
+        gtk_grid_attach(GTK_GRID(grid),gtk_button_new_from_icon_name("gtk-delete",GTK_ICON_SIZE_BUTTON),3,i,1,1);
+        gtk_button_set_label(GTK_BUTTON(gtk_grid_get_child_at(GTK_GRID(grid),1,i)),_("Edit"));
+        gtk_button_set_label(GTK_BUTTON(gtk_grid_get_child_at(GTK_GRID(grid),2,i)),_("Properties"));
+        gtk_button_set_label(GTK_BUTTON(gtk_grid_get_child_at(GTK_GRID(grid),3,i)),_("Delete"));
+        #endif // _WIN32
+        g_signal_connect (gtk_grid_get_child_at(GTK_GRID(grid),3,i),"clicked", G_CALLBACK(deleteGameConfiguration),data);
+        g_signal_connect (gtk_grid_get_child_at(GTK_GRID(grid),2,i),"clicked", G_CALLBACK(viewGameConfiguration),data);
+        g_signal_connect (gtk_grid_get_child_at(GTK_GRID(grid),1,i),"clicked", G_CALLBACK(editGameConfiguration),data);
+
+        gtk_widget_set_tooltip_markup(gtk_grid_get_child_at(GTK_GRID(grid),2,i),_("View the details of the game configuration"));
+        gtk_widget_set_tooltip_markup(gtk_grid_get_child_at(GTK_GRID(grid),3,i),_("Delete the game configuration"));
     }
 
     /* Add the different containers */
@@ -243,15 +276,485 @@ void displayGameConfiguration(GtkWidget *window_game)
     gtk_container_add(GTK_CONTAINER(viewport),grid);
     gtk_container_add(GTK_CONTAINER(window_scrolled),viewport);
 
+    gtk_widget_show_all(window_scrolled);
+
     closeListGameConfig(ptr_list_config);
 }
 
 /*!
- * \fn void void deleteDisplayGameConfiguration(GtkWidget *window_game)
+ * \fn void deleteDisplayGameConfiguration(GtkWidget *window_game)
  *  Delete the display a all the game configuration on the window
- * \param[in] window_game the window of game configuration
+ * \param[in] data the globalData
  */
-void deleteDisplayGameConfiguration(GtkWidget *window_game)
+void deleteDisplayGameConfiguration(globalData *data)
 {
-    gtk_widget_destroy(gtk_bin_get_child(GTK_BIN(gtk_grid_get_child_at(GTK_GRID(gtk_bin_get_child(GTK_BIN(window_game))),0,1))));
+    GtkWidget *window_scrolled = GTK_WIDGET(gtk_builder_get_object(data->ptr_builder,"scrolled_window_game_configuration_preferences"));
+    if (!window_scrolled)
+        g_critical(_("Widget scrolled_window_game_configuration_preferences is missing in file csuper-gui.glade."));
+
+    gtk_widget_destroy(gtk_bin_get_child(GTK_BIN(window_scrolled)));
+}
+
+/*!
+ * \fn void deleteDisplayGameConfiguration(GtkWidget *window_game)
+ *  Update the display a all the game configuration on the window
+ * \param[in] data the globalData
+ */
+void updateDisplayGameConfiguration(globalData *data)
+{
+    deleteDisplayGameConfiguration(data);
+    displayGameConfiguration(data);
+}
+
+/*!
+ * \fn G_MODULE_EXPORT void deleteGameConfiguration(GtkWidget *widget, gpointer data)
+ *  Delete a game configuration
+ * \param[in] widget the widget which send the interrupt
+ * \param[in] data the globalData
+ */
+G_MODULE_EXPORT void deleteGameConfiguration(GtkWidget *widget, gpointer data)
+{
+    globalData *user_data = (globalData*) data;
+    gint i=0;
+
+    /* Get the button clicked*/
+    GtkWidget *window_scrolled = GTK_WIDGET(gtk_builder_get_object(user_data->ptr_builder,"scrolled_window_game_configuration_preferences"));
+    if (!window_scrolled)
+        g_critical(_("Widget scrolled_window_game_configuration_preferences is missing in file csuper-gui.glade."));
+    GtkGrid *grid = GTK_GRID(gtk_bin_get_child(GTK_BIN(gtk_bin_get_child(GTK_BIN(window_scrolled)))));
+    while (1)
+    {
+        if (gtk_grid_get_child_at(GTK_GRID(grid),3,i) == widget)
+            break;
+        i++;
+    }
+
+    /* Manage the game configurations */
+    list_game_config *ptr_list_config;
+    char home_path[SIZE_MAX_FILE_NAME]="";
+    #ifndef PORTABLE
+    readHomePathSlash(home_path);
+    #endif // PORTABLE
+    ptr_list_config = readConfigListFile(home_path);
+    removeConfigListFile(i,ptr_list_config,home_path);
+    closeListGameConfig(ptr_list_config);
+
+    updateDisplayCurrentGameConfiguration(user_data,i,TRUE);
+    updateDisplayGameConfiguration(user_data);
+}
+
+/*!
+ * \fn G_MODULE_EXPORT void editGameConfiguration(GtkWidget *widget, gpointer data)
+ *  Edit a game configuration
+ * \param[in] widget the widget which send the interrupt
+ * \param[in] data the globalData
+ */
+G_MODULE_EXPORT void editGameConfiguration(GtkWidget *widget, gpointer data)
+{
+    globalData *user_data = (globalData*) data;
+    gint i=0;
+    game_config config;
+    game_config *ptr_config;
+    list_game_config *ptr_list_config;
+    char home_path[SIZE_MAX_FILE_NAME]="";
+
+    /* Get the button clicked*/
+    GtkWidget *window_scrolled = GTK_WIDGET(gtk_builder_get_object(user_data->ptr_builder,"scrolled_window_game_configuration_preferences"));
+    if (!window_scrolled)
+        g_critical(_("Widget scrolled_window_game_configuration_preferences is missing in file csuper-gui.glade."));
+    GtkGrid *grid = GTK_GRID(gtk_bin_get_child(GTK_BIN(gtk_bin_get_child(GTK_BIN(window_scrolled)))));
+    while (1)
+    {
+        if (gtk_grid_get_child_at(GTK_GRID(grid),1,i) == widget)
+            break;
+        i++;
+    }
+
+    #ifndef PORTABLE
+    readHomePathSlash(home_path);
+    #endif // PORTABLE
+    ptr_list_config = readConfigListFile(home_path);
+    readConfigFile(i,ptr_list_config,&config,home_path);
+
+    changeNewGameConfigurationDialog(user_data,config);
+    ptr_config = newGameConfiguration(user_data);
+    if(ptr_config != NULL)
+    {
+        removeConfigListFile(i,ptr_list_config,home_path);
+        newConfigFile(*ptr_config,home_path);
+        free(ptr_config);
+        updateDisplayGameConfiguration(user_data);
+        updateDisplayCurrentGameConfiguration(user_data,i,TRUE);
+    }
+
+    closeListGameConfig(ptr_list_config);
+}
+
+/*!
+ * \fn G_MODULE_EXPORT viewGameConfiguration(GtkWidget *widget, gpointer data)
+ *  Display the game configuration
+ * \param[in] widget the widget which send the interrupt
+ * \param[in] data the globalData
+ */
+G_MODULE_EXPORT void viewGameConfiguration(GtkWidget *widget, gpointer data)
+{
+    globalData *user_data = (globalData*) data;
+    gint i=0;
+
+    /* Get the button clicked*/
+    GtkWidget *window_scrolled = GTK_WIDGET(gtk_builder_get_object(user_data->ptr_builder,"scrolled_window_game_configuration_preferences"));
+    if (!window_scrolled)
+        g_critical(_("Widget scrolled_window_game_configuration_preferences is missing in file csuper-gui.glade."));
+    GtkGrid *grid = GTK_GRID(gtk_bin_get_child(GTK_BIN(gtk_bin_get_child(GTK_BIN(window_scrolled)))));
+    while (1)
+    {
+        if (gtk_grid_get_child_at(GTK_GRID(grid),2,i) == widget)
+            break;
+        i++;
+    }
+
+    updateDisplayCurrentGameConfiguration(user_data,i,FALSE);
+}
+
+/*!
+ * \fn void updateDisplayCurrentGameConfiguration(globalData *data , gint index, gboolean clear)
+ *  Update the current game configuration label
+ * \param[in] data the globalData
+ * \param[in] index the index of the game configuration, -1 if you want to clear the label
+ * \param[in] clear if TRUE the label is reset if the index is the same than the last call to the function otherwise the label is update
+ */
+void updateDisplayCurrentGameConfiguration(globalData *data , gint index, gboolean clear)
+{
+    globalData *user_data = (globalData*) data;
+    static gint last_index=-1;char *yes=_("yes");
+    char *no=_("no");
+
+    if (last_index == index && clear)
+        index = -1;
+
+    if (!clear)
+        last_index = index;
+    else
+    {
+        if (index < last_index)
+            last_index--;
+    }
+
+    /* Get the labels */
+    GtkWidget *label_details = GTK_WIDGET(gtk_builder_get_object(user_data->ptr_builder,"game_configuration_preferences_label_details"));
+    if (!label_details)
+        g_critical(_("Widget game_configuration_preferences_label_details is missing in file csuper-gui.glade."));
+    GtkWidget *label_name = GTK_WIDGET(gtk_builder_get_object(user_data->ptr_builder,"game_configuration_preferences_label_name"));
+    if (!label_name)
+        g_critical(_("Widget game_configuration_preferences_label_name is missing in file csuper-gui.glade."));
+
+    if (index >= 0 && !clear)
+    {
+        /* Manage the game configurations */
+        game_config config;
+        list_game_config *ptr_list_config;
+        char home_path[SIZE_MAX_FILE_NAME]="";
+        #ifndef PORTABLE
+        readHomePathSlash(home_path);
+        #endif // PORTABLE
+        ptr_list_config = readConfigListFile(home_path);
+        readConfigFile(index,ptr_list_config,&config,home_path);
+
+        /* Print the game config */
+        #ifdef _WIN32
+        if (config.nb_max != INFINITY)
+        {
+            gtk_label_set_text(GTK_LABEL(label_details),g_strdup_printf(_("Use a maximum score : %s\nNumber of points maximum/minimum : %.3f\n"
+                "Initial score : %.3f\nNumber of decimal place : %d\nThe first has the highest score : %s\n"
+                "Game in turn by turn : %s\nUse a distributor : %s"),integerToYesNo(config.max,yes,no),config.nb_max
+                ,config.begin_score,config.decimal_place,integerToYesNo(config.first_way,yes,no),
+                integerToYesNo(config.turn_by_turn,yes,no),integerToYesNo(config.use_distributor,yes,no)));
+        }
+        else
+        {
+            gtk_label_set_text(GTK_LABEL(label_details),g_strdup_printf(_("Use a maximum score : %s\nNumber of points maximum/minimum : inf\n"
+                "Initial score : %.3f\nNumber of decimal place : %d\nThe first has the highest score : %s\n"
+                "Game in turn by turn : %s\nUse a distributor : %s"),integerToYesNo(config.max,yes,no)
+                ,config.begin_score,config.decimal_place,integerToYesNo(config.first_way,yes,no),
+                integerToYesNo(config.turn_by_turn,yes,no),integerToYesNo(config.use_distributor,yes,no)));
+        }
+        #else
+        gtk_label_set_text(GTK_LABEL(label_details),g_strdup_printf(_("Use a maximum score : %s\nNumber of points maximum/minimum : %.3f\n"
+            "Initial score : %.3f\nNumber of decimal place : %d\nThe first has the highest score : %s\n"
+            "Game in turn by turn : %s\nUse a distributor : %s"),integerToYesNo(config.max,yes,no),config.nb_max
+            ,config.begin_score,config.decimal_place,integerToYesNo(config.first_way,yes,no),
+            integerToYesNo(config.turn_by_turn,yes,no),integerToYesNo(config.use_distributor,yes,no)));
+        #endif // _win32
+        gtk_label_set_text(GTK_LABEL(label_name),g_strdup_printf(_("%s"),config.name));
+
+        closeListGameConfig(ptr_list_config);
+    }
+    else
+    {
+        if (index < 0)
+        {
+            gtk_label_set_text(GTK_LABEL(label_details),g_strdup_printf(_("You have no game configuration selected.")));
+            gtk_label_set_text(GTK_LABEL(label_name),"");
+        }
+    }
+
+    gtk_widget_show(label_details);
+}
+
+/*!
+ * \fn G_MODULE_EXPORT addGameConfiguration(GtkWidget *widget, gpointer data)
+ *  Add a game configuration
+ * \param[in] widget the widget which send the interrupt
+ * \param[in] data the globalData
+ */
+G_MODULE_EXPORT void addGameConfiguration(GtkWidget *widget, gpointer data)
+{
+    globalData *user_data = (globalData*) data;
+    game_config *ptr_config = NULL;
+    char home_path[SIZE_MAX_FILE_NAME]="";
+    game_config config = {0,1,1,0,-1,-1,"",0};
+
+    changeNewGameConfigurationDialog(user_data,config);
+    ptr_config=newGameConfiguration(user_data);
+    if (ptr_config == NULL)
+        return;
+
+    #ifndef PORTABLE
+    readHomePathSlash(home_path);
+    #endif // PORTABLE
+
+    newConfigFile(*ptr_config,home_path);
+
+    free(ptr_config);
+
+    updateDisplayGameConfiguration(user_data);
+}
+
+
+/*!
+ * \fn game_config *newGameConfiguration(globalData *data)
+ *  Display the window of creating a game configuration and return the game configuration created
+ * \param[in] data the globalData
+ */
+game_config *newGameConfiguration(globalData *data)
+{
+    game_config config;
+    game_config *ptr_config = NULL;
+    gchar *buffer;
+
+    /* Get the dialog box */
+    GtkWidget *dialog = GTK_WIDGET(gtk_builder_get_object(data->ptr_builder,"new_game_configuration_dialog"));
+    if (!dialog)
+        g_critical(_("Widget new_game_configuration_dialog is missing in file csuper-gui.glade."));
+
+    /* Get the grid */
+    GtkWidget *grid = GTK_WIDGET(gtk_builder_get_object(data->ptr_builder,"new_game_configuration_dialog_grid"));
+    if (!grid)
+        g_critical(_("Widget new_game_configuration_dialog_grid is missing in file csuper-gui.glade."));
+
+    /* Define the size max of the game configuration name */
+    gtk_entry_set_max_length(GTK_ENTRY(gtk_grid_get_child_at(GTK_GRID(grid),1,0)),SIZE_MAX_NAME);
+
+    gtk_widget_show_all(dialog);
+
+    switch (gtk_dialog_run(GTK_DIALOG(dialog)))
+	{
+		case GTK_RESPONSE_ACCEPT:
+		{
+		    /* Name */
+		    strcpy(config.name,gtk_entry_get_text(GTK_ENTRY(gtk_grid_get_child_at(GTK_GRID(grid),1,0))));
+		    if (strcmp(config.name,"") == 0)
+                break;
+
+		    /* Max / min */
+		    buffer = gtk_combo_box_text_get_active_text(GTK_COMBO_BOX_TEXT(gtk_grid_get_child_at(GTK_GRID(grid),1,1)));
+		    if (buffer == NULL)
+                break;
+            else
+            {
+                if (strcmp(buffer,_("Maximum")) == 0)
+                    config.max = 1;
+                if (strcmp(buffer,_("Minimum")) == 0)
+                    config.max = 0;
+                if (strcmp(buffer,_("Neither")) == 0)
+                {
+                    config.max = 1;
+                    config.nb_max = INFINITY;
+                }
+                else
+                {
+                    config.nb_max = gtk_spin_button_get_value(GTK_SPIN_BUTTON(gtk_grid_get_child_at(GTK_GRID(grid),1,2)));
+                }
+                g_free(buffer);
+            }
+
+            /* Initial score */
+            config.begin_score = gtk_spin_button_get_value(GTK_SPIN_BUTTON(gtk_grid_get_child_at(GTK_GRID(grid),1,3)));
+
+            /* First way */
+            if (gtk_switch_get_active(GTK_SWITCH(gtk_grid_get_child_at(GTK_GRID(grid),1,4))))
+                config.first_way = 1;
+            else
+                config.first_way = -1;
+
+            /* Turn by turn */
+            if (gtk_switch_get_active(GTK_SWITCH(gtk_grid_get_child_at(GTK_GRID(grid),1,5))))
+                config.turn_by_turn = 1;
+            else
+                config.turn_by_turn = 0;
+
+            /* Distributor */
+            if (gtk_switch_get_active(GTK_SWITCH(gtk_grid_get_child_at(GTK_GRID(grid),1,6))))
+                config.use_distributor = 1;
+            else
+                config.use_distributor = 0;
+
+            /* Decimal place*/
+		    buffer = gtk_combo_box_text_get_active_text(GTK_COMBO_BOX_TEXT(gtk_grid_get_child_at(GTK_GRID(grid),1,7)));
+		    if (buffer == NULL)
+                break;
+            else
+            {
+                if (strcmp(buffer,_("0")) == 0)
+                    config.decimal_place = 0;
+                if (strcmp(buffer,_("1")) == 0)
+                    config.decimal_place = 1;
+                if (strcmp(buffer,_("2")) == 0)
+                    config.decimal_place = 2;
+                if (strcmp(buffer,_("3")) == 0)
+                    config.decimal_place = 3;
+
+                g_free(buffer);
+            }
+
+            ptr_config = myAlloc(sizeof(game_config));
+            *ptr_config = config;
+
+			break;
+		}
+		case GTK_RESPONSE_CANCEL:
+        {
+            return ptr_config;
+            break;
+        }
+		default:
+		    return ptr_config;
+			break;
+	}
+
+    if (ptr_config == NULL)
+        newGameConfigurationError(data);
+
+	return ptr_config;
+}
+
+/*!
+ * \fn void newGameConfigurationError(globalData *data)
+ *  Display a dialog box which said that there is a problem on the new game configuration.
+ * \param[in] data the globalData
+ */
+void newGameConfigurationError(globalData *data)
+{
+    GtkWidget *window_error = GTK_WIDGET(gtk_builder_get_object(data->ptr_builder,"message_dialog_error_add_game_configuration"));
+    if (!window_error)
+        g_critical(_("Widget message_dialog_error_add_game_configuration is missing in file csuper-gui.glade."));
+
+    gtk_dialog_run (GTK_DIALOG (window_error));
+    gtk_widget_hide (window_error);
+}
+
+/*!
+ * \fn G_MODULE_EXPORT noMaxMinNewGameConfiguration(GtkWidget *widget, gpointer data)
+ *  Change text editable propriety of the max/min value
+ * \param[in] widget the widget which send the interrupt
+ * \param[in] data the globalData
+ */
+G_MODULE_EXPORT void noMaxMinNewGameConfiguration(GtkWidget *widget, gpointer data)
+{
+    globalData *user_data = (globalData*) data;
+    gchar *buffer;
+
+    GtkWidget *max = GTK_WIDGET(gtk_builder_get_object(user_data->ptr_builder,"new_game_configuration_spin_nb_max"));
+    if (!max)
+        g_critical(_("Widget new_game_configuration_spin_nb_max is missing in file csuper-gui.glade."));
+
+    buffer = gtk_combo_box_text_get_active_text(GTK_COMBO_BOX_TEXT(widget));
+    if (buffer == NULL)
+        return;
+
+    if (strcmp(buffer,_("Neither")) == 0)
+            gtk_editable_set_editable(GTK_EDITABLE(max),FALSE);
+    else
+        gtk_editable_set_editable(GTK_EDITABLE(max),TRUE);
+
+    g_free(buffer);
+}
+
+/*!
+ * \fn void changeNewGameConfigurationDialog(globalData *data,game_config config)
+ *  Enter the game configuration into the new game configuration dialog
+ * \param[in] data the globalData
+ * \param[in] config a game config
+ */
+void changeNewGameConfigurationDialog(globalData *data,game_config config)
+{
+    GtkWidget *grid = GTK_WIDGET(gtk_builder_get_object(data->ptr_builder,"new_game_configuration_dialog_grid"));
+    if (!grid)
+        g_critical(_("Widget new_game_configuration_dialog_grid is missing in file csuper-gui.glade."));
+
+    gtk_entry_set_text(GTK_ENTRY(gtk_grid_get_child_at(GTK_GRID(grid),1,0)),config.name);
+
+    if (config.max == 0)
+        gtk_combo_box_set_active_id(GTK_COMBO_BOX(gtk_grid_get_child_at(GTK_GRID(grid),1,1)),"min");
+    else
+    {
+        if (config.max == 1)
+        {
+            if (config.nb_max == INFINITY)
+                gtk_combo_box_set_active_id(GTK_COMBO_BOX(gtk_grid_get_child_at(GTK_GRID(grid),1,1)),"neither");
+            else
+                gtk_combo_box_set_active_id(GTK_COMBO_BOX(gtk_grid_get_child_at(GTK_GRID(grid),1,1)),"max");
+        }
+        else
+            gtk_combo_box_set_active(GTK_COMBO_BOX(gtk_grid_get_child_at(GTK_GRID(grid),1,1)),-1);
+    }
+
+
+    gtk_spin_button_set_value(GTK_SPIN_BUTTON(gtk_grid_get_child_at(GTK_GRID(grid),1,2)),config.nb_max);
+
+    gtk_spin_button_set_value(GTK_SPIN_BUTTON(gtk_grid_get_child_at(GTK_GRID(grid),1,3)),config.begin_score);
+
+    if(config.first_way == 1)
+        gtk_switch_set_active(GTK_SWITCH(gtk_grid_get_child_at(GTK_GRID(grid),1,4)),TRUE);
+    else
+        gtk_switch_set_active(GTK_SWITCH(gtk_grid_get_child_at(GTK_GRID(grid),1,4)),FALSE);
+
+    if(config.turn_by_turn == 1)
+        gtk_switch_set_active(GTK_SWITCH(gtk_grid_get_child_at(GTK_GRID(grid),1,5)),TRUE);
+    else
+        gtk_switch_set_active(GTK_SWITCH(gtk_grid_get_child_at(GTK_GRID(grid),1,5)),FALSE);
+
+    if(config.use_distributor == 1)
+        gtk_switch_set_active(GTK_SWITCH(gtk_grid_get_child_at(GTK_GRID(grid),1,6)),TRUE);
+    else
+        gtk_switch_set_active(GTK_SWITCH(gtk_grid_get_child_at(GTK_GRID(grid),1,6)),FALSE);
+
+    switch(config.decimal_place)
+    {
+    case 0:
+        gtk_combo_box_set_active_id(GTK_COMBO_BOX(gtk_grid_get_child_at(GTK_GRID(grid),1,7)),"0");
+        break;
+    case 1:
+        gtk_combo_box_set_active_id(GTK_COMBO_BOX(gtk_grid_get_child_at(GTK_GRID(grid),1,7)),"1");
+        break;
+    case 2:
+        gtk_combo_box_set_active_id(GTK_COMBO_BOX(gtk_grid_get_child_at(GTK_GRID(grid),1,7)),"2");
+        break;
+    case 3:
+        gtk_combo_box_set_active_id(GTK_COMBO_BOX(gtk_grid_get_child_at(GTK_GRID(grid),1,7)),"3");
+        break;
+    default:
+        gtk_combo_box_set_active(GTK_COMBO_BOX(gtk_grid_get_child_at(GTK_GRID(grid),1,7)),-1);
+        break;
+    }
 }